Implement handling of sink cancellation
Summary:
The added unit test revealed that sink cancellation was not previously handled. If the input async generator is destroyed, it is possible that we need to send a sink cancellation signal to the client.

Note that `SCOPE_EXIT` is insufficient - it will never run if the generator is never co_awaited.

Reviewed By: robertroeser, sazonovkirill

Differential Revision: D84009581

fbshipit-source-id: b6952b9cf037c60e574db22eb8b9611a63d3b81d