Fix readonly
Summary:
The function type in the constraint type needs to be marked readonly return if the method call is inside of a readonly expression: as in `readonly $x->m()`. in this case, when we determine the function type for `m`, it will be able to be readonly, since a readonly return function is a supertype of a non-readonly return function.

If the receiver is readonly `$x = readonly ...; $x->m()`, then the function in the constraint should be maked as readonly this, because readonly this functions are subtypes of non-readonly this functions, and we need to require m to be readonly.

If the argument is readonly, we need to record that in the constraint's function type parameter, so that if the method we are calling turns out to readonly, that will be allowed, since a readonly parameter function is a subtype of one without.

The way this works is brittle and confusing because we are relying on the parser to introduce ReadonlyExpr wrappers immediately around arguments and obj_gets that are readonly, which we look for here.

Reviewed By: andrewjkennedy

Differential Revision: D84513163

fbshipit-source-id: aa808d9bdcfce8b5f20f210a2dc5963af228025a