Only increase rank when we have universally quantified supertype
Summary:
When subtyping universally quantified function types we have two case should be dealt with in the following order

```
T <: forall U. V
```

and

```
forall U. V <: T
```

In the first case increment the rank and bind all generics. In the second case we can just instantiate the generics to get back an unquantified type.

If a universally quantified type appears on the lhs and a universally quantified type was on the rhs, it will be instantiated at the correct rank so we don't need to worry about the following case explicitly:

```
forall U. V <: forall S. T
```

Currently the subtyping implementation does these two steps simultaneously but the condition for incrementing rank is that _either_ type is quantified rather than just the supertype. This diff corrects the issue.

Reviewed By: andrewjkennedy

Differential Revision: D85059488

fbshipit-source-id: 43cb698bed738cc597861f9cac573364f3d00482