small refactor
Summary: Lift the checking of the receiver expression so that we can base our decision on whether to use constraints on how solved its type is in the future. We duplicated the suffix in the if, instead of sharing after the if, because it's not that much code, and there would have to be many values returned from the if branches into a let to make this work.

Reviewed By: andrewjkennedy

Differential Revision: D84615076

fbshipit-source-id: ec8802ca85c0e90fcf9410b7d34dafc0799c8e60