Check that type splat is a tuple for method parameters
Summary:
Earlier diff D74829336 checked that in a parameter declaration of the form `...t` the `t` is a tuple type. But it did this only for top-level functions, not methods (static or instance). We rectify that in this diff.

(Note that we have multiple subtype errors for the same declaration, which is unfortunate, and could be fixed in a separate change.)

Differential Revision: D85052591

fbshipit-source-id: 5d49c8a742e0e515a507df6d6dd6839b86d716c1