keep ClassGraph::SerdeState containers reserved across invocations
Summary:
Some workloads such as the AnalyzeConstants job write/read many class graphs, and the SerdeState containers get built up and destroyed constantly despite them fundamentally being vectors that hold either StringData*'s or ints as data, none of which is owned by them.

This diff changes the SerdeState management to intentionally leak the containers across serialization/deserialization jobs. The memory leak is bounded by the size of the largest class graph's usage, and in practice this reduces the memory traffic associated with SerDe'ing class graphs significantly.

Because we're now allocating the thread_local state inline (instead of via a pointer), I had to move some forward-declarations back in ClassGraph so we get a statically known size to allocate.

Reviewed By: ricklavoie

Differential Revision: D84212344

fbshipit-source-id: 30de1091f3950c20fe324020577e5285c07e449a