Add methods to look up note by name or by type
Summary:
This rounds out my changes to Elf, exposing easy to use iterators for type and name. Because Notes are not necessarily stored uniquely, I've removed the complexity of it being a note in a section or a segment and instead expose generic Note iteration.

I think this is valuable because a user often just wants a note of a given type regardless of it being in section or segment.

Reviewed By: dmpots

Differential Revision: D84076528

fbshipit-source-id: 1bdb287ceb7a14153e22e3cf5afb8efae5b518a5