strip position from class names and package information
Summary:
When comparing classes to detect if they have "minor" changes, the positions of the class ids were compared as well: as a result, adding a newline before a class definition would result in retypechecking its fanout.  This diff refines the pos_id check: it preserves the comparison of the id and of the filename (as moving a definition to a new file should trigger fanout), but it makes it agnostic to the position inside the file.

Additionally, when detecting "major" changes the position of package information was not stripped, so modifying the position of a PackageOverride would trigger retypechecking all the definitions in the file.

[ also remove `class_big_diff` which is unused. ]

Reviewed By: CatherineGasnier, andrewjkennedy

Differential Revision: D85049227

fbshipit-source-id: 222c57a749ae0a8ae67c69ac528e897d9c33d3fc