Do not lose bounds on generic parameters under refinement
Summary:
When we do a type-test such as `$x as C<_>`, we generate fresh type parameters for the wildcards, and then attempt to "optimize" this by pushing them to their bounds. For example, if `C` is defined as `class C<+T>` and `$x` is of type `mixed`, we first generate a fresh type parameter `T#1` to produce `C<T#1>` but then optimize it by pushing it to `mixed`, because `T` is covariant.

Unfortunately we fail to apply this "pushing" to occurrences of the fresh generic in the bounds of other generics. Here's an example:
```
function expect_mixed(mixed $_):void { }

interface I { }

abstract class VC { }

interface IA<T as I> { }

class B<TVC as VC> implements I {
  public function get():TVC {
    throw new Exception("A");
  }
}

function expect_vc(VC $_):void { }

interface IB<-TVC as VC, T as B<TVC>> extends IA<T> {
  public function get_t(): T;
}

class C<T as I> {
  public function testit(T $x, IA<T> $m):void {
    $m as IB<_,_>;
    // This used to be rejected under everything_sdt=true
    expect_mixed($m);
    $x = $m->get_t();
    $y = $x->get();
    // This used to be rejected ALSO under everything_sdt=false
    // It should be accepted
    expect_vc($y);
  }
}
```
At the refinement `$m as IB<_,_>` we first generate fresh generics, to produce `IB<TVC#1, T#1>` with bounds `TVC#1 as VC` and `T#1 as B<TVC#1>`. We now optimize according to the variance to get `IB<nothing,T#1>`, and end up eliminating `TVC#1` from the type parameter environment completely. But unfortunately it's there in the bound of `T#1`.

The fix is to apply the pushing (a substitution) to the bounds in the type parameter environment, so we get `T#1 as B<nothing>`.

The manifestation of this bug is particularly bad under pessimisation, because we can end up losing the `supportdyn<mixed>` implicit bound on generics, leading to errors such as `mixed is not a subtype of dynamic`.

Reviewed By: enetsee

Differential Revision: D85261626

fbshipit-source-id: 28434cd782fd21464ca22b2d0a2b5e8325b79159