Special case typing of top-level function- and static method pointers
Summary:
The treatment of function pointers to top-level function and static methods is baked into the lowering of expression trees. The new typing as first-class polymorphic functions introduces two issues:

1) Inference behavior is changed meaning we need hints in many places we didn't before

2) Type checking performance degrades for very complex expression tree blocks

This diff adds a flag to the funciton pointer AST nodes and modifies typing so that we instantiate polymorphic function types when they were generated during lowering.

Reviewed By: andrewjkennedy

Differential Revision: D85429843

fbshipit-source-id: 9772b6fb3c31a6b5f25dc8cb72df7db5deeecfde