Add warning messages to usages of `hasattr(obj, thrift_spec)`
Summary:
**Background:**
During the SEV reviews for migration diffs, a common issue identified was the use of the internal attribute `thrift_spec` from thrift-py-deprecated, which does not exist in thrift-python.

**Summary:**
This diff introduces warning messages wherever `hasattr(obj, thrift_spec)` is used. The goal is to alert developers to update any logic that relies on thrift_spec, ensuring compatibility with thrift-python and preventing future migration issues.

Reviewed By: jhshi

Differential Revision: D85376645

fbshipit-source-id: b8bdfb44a7ba14312cfb76179f95c50f0cf809b1