Wrap Watchman into a generic Watcher Plugin
Summary:
We want to add support for Eden and rather than having both an ext_watchman + ext_eden, we're developing a more generic ext_watcher that encapsulates the idea of file watching, with the backend being used being configurable.

This takes a WatcherOption configuration and then dynamically generates an appropriate Watchman query from it.

This is a work in progress and should not yet be used.

Some current issues:
- No file for the typechecker yet (intentional)
- Output format doesn't yet match the specification
- Only supports watchman (so far...)

Differential Revision: D85023217

fbshipit-source-id: b6dec22645478aaf64ae4a2acec11b565a5cdc19