Added a new thrift Adapter called PaddedBinaryAdapter.
Summary:
This change adds a new thrift adapter called PaddedBinaryAdapter that adds padding to a binary field and helps serialized and deserialized that padding.

Users MUST serialize and deserialize their data as a PaddedBinaryData in order to make sure:
1. Padding is added to the field
2. Padded and any relevant metadata is correctly removed and the actual user data is returned back to the user.

This adapter modifies the user data blob and adds additional information to it to make the actual user data padded. The data buffer is then serialized and deserialized as a native thrift binary field.

The modified user data that is encoded looks like:
[8 bytes magic][4 bytes padding size field][padding of n bytes][user data]

If the client sends a simple binary (not encoded using PaddedBinaryData), it can be correctly decoded using PaddedBinaryAdapter.
If the client sends data encoded using PaddedBinaryData, the server can decode it as a simple binary, but  it would NOT remove the extra information prepended to the user data.

Users must upgrade the server side first to use this, and then the client side in order to be able to parse the padding and remove it correctly on the server side.

Reviewed By: robertroeser

Differential Revision: D85457710

fbshipit-source-id: 07adf7c9b5ce8ae44779b5d34600a48829eef834