Modify QueueAppender::insert to allow not packing the new iobuf with the current iobuf.
Summary:
IOBufQueue today supports a way to insert iobufs without packing, but QueueAppender doesn't. This change is to make QueueAppender support insertion of iobufs into the underlying queue without packing them.

We need it in thrift for our supporting alignment for zero copy, but it also helps a broader audience that might want to keep the iobuf separate from the existing data. Given that there is existing support in IOBufQueue for doing this, I think having that support with QueueAppender would be great, since that's a cursor interface for IOBufQueue.

Reviewed By: robertroeser

Differential Revision: D85097847

fbshipit-source-id: f4420c53a3ebcaa0c031cae1128aa23817f2b940