Debugger - stepIn should set stack depth
Summary:
The issue - users report that `stepIn` after a breakpoint hit is going
to the wrong function.

My first guess: phpDebgugerOpcodeHook gets called on processing the opcode after
stepping in. It calls `auto curStackDisp =
getStackDisposition(req_data.getDebuggerFlowDepth());` but we never set the
flow depth. Note that this is different from stepOut and stepNext. Manually
testing suggests this helps.

Reviewed By: slinakm

Differential Revision: D85604472

fbshipit-source-id: 242abfe91c52c31dfa53cf90a1b43e275c8f2cfd