Fix bug in localizing mutually recursive case types
Summary:
When we localize a case type, we localize its upper bound. If the user provides multiple upper bounds, the type that is localized is an intersection and we localize the intersection. When we localize an intersection, we simplify it, but [do not pass the ety_env to the simplification logic](https://www.internalfb.com/code/fbsource/[a63d986d58ab]/fbcode/hphp/hack/src/typing/typing_phase.ml?lines=561). When the simplification logic [does subtyping](https://www.internalfb.com/code/fbsource/[a41cc2549467]/fbcode/hphp/hack/src/typing/typing_intersection.ml?lines=251-256), we will open the case type and localize the variants. This can result in infinite recursion as we bounce back between localization and subtyping since we do not propagate the ety_env.

The solution is to add the case type to the expansions for cycle checking when localizing the variants.

This solution was added in D49560162 (see last part of description) but was lost in D60669396 (perhaps b/c it was assumed the recursion tracker the diff added made the expansions addition obsolete).

Reviewed By: dlreeves

Differential Revision: D85703967

fbshipit-source-id: 1b80ddf4f2b0c3afaae203855650708da08fa785