bidi mustache
Summary:
Updates the mustache templates to implement bidirectional streaming server side. The logic is verbose because in the case of a bidirectional method both `#function:bidirectional_stream?` and `#function:stream?` (the same for sink) return true.

AI helped with this and introduced formatting changes:
- adding semicolons to the end of lines
- adding a line break for large `Result` types

I was originally going to revert them, but I've grown to like them when looking at the generated code. I'm not completely sold on these format changes, so I'll remove them if anyone doesn't like them.

Reviewed By: dtolnay

Differential Revision: D85113081

fbshipit-source-id: 4a0aa26e1f99a2e76b4838c1946f03c822e7e8b7