fix broken links and anchors
Summary:
### Changes

Fix up links and anchors to make the md-files compatible, and update config and conventions to be in line with the [3.9.x upgrade](https://docusaurus.io/docs/api/docusaurus-config#onBrokenLinks).

### Note

Some links to `/docs/apis/*` will log as warnings until we auto-materialise Hack API and HSL references on the new doc site.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D85745533

fbshipit-source-id: 09b5178701481b2d100dd9331f30948362fa7c79