Pessimise parameters for single-parameter enforced-return functions and non-overridable methods
Summary:
When checking calls to functions that support dynamic calling (the default under implicit pessimisation), we propagate like-types from argument to return types, in effect implementing the call as a call through an intersection of function types. For example:
```
function foo(vec<int> $x): vec<string>;
```
is treated as having a type `(function(vec<int>):vec<string>) & (function(dynamic):dynamic)` (actually, represented as `supportdyn<(function(vec<int>):vec<string>)>`) so calls with an argument of type `~vec<int>` will produce a return type of `~vec<string>`.

Unfortunately this applies to all functions, even those whose return  is *known* to be enforced, regardless of whether the function is called dynamically, or with an argument of like-type. For example, calling a function
```
function foo(int):string;
```
with an argument of type `~int` will produce a result of type `~string`, even though `foo` will always enforce its return type.

To avoid this, the function can be given a stronger type: `(function(~int):string)`, instead of `supportdyn<(function(int):string)>`, so the result is a `string` regardless of whether the argument has type `int`, or `dynamic`, or `~int`.

Because override doesn't preserve enforcement, we can't apply this policy uniformly, unless we are prepared to "poison" signatures up hierarchies. So for now we apply it to "non-overridable" functions: top-level functions, non-trait private methods, final methods, and methods in final classes.

UPDATE: sowens points out that this change is unsound for multiple argument functions, in the presence of types that do not support dynamic. We have a proposed solution which is for the second pass of typechecking a function or method use *like types* instead of *dynamic* for the parameter types, i.e. matching the effect of the pessimisation scheme described here. Until that rolls out, we restrict this scheme to *single parameter* functions.

Differential Revision: D82813368

fbshipit-source-id: 12e2f1ce314d3c3136d62a9520b6c738e8c61a1c