Omit like types in decl hover
Summary:
Currently we show like types (from pessimisation) even for "decl" types such as that shown when hovering over calls to functions. This is not very helpful, especially now that we're employing more clever pessimisation e.g. putting likes on parameter types in order to prevent like-propagation through to enforced returns. For this case, let's omit the likes. e.g. show
 {F1983163734}
and not
 {F1983170217}

Reviewed By: madgen

Differential Revision: D85961760

fbshipit-source-id: c031eb41aba8f700ee0e07430572c72aa1362765