AArch64: Improve code quality during type checking on Arm targets
When checking a type matches we often need to compare against an
arbitrary bit mask (usually fits into a single byte). However,
the tst instruction on AArch64 only supports masks of contiguous
bits so usually the mask cannot be encoded into the tst
instruction. Instead, when compiling for Arm we should prefer
using the cmp instruction as this can take any 8-bit immediate.