test helper converts to float32
Summary:
Add a testing helper usable in customer code to convert to `float32`. This includes collections of numbers, like common Python container types and thrift-python concepts.

It is anticipated that most usage will be immutable thrift-python, so mutable thrift containers are not yet implemented. They are non-trivial to implement to the required `to_thrift_list` helper when being used as a thrift field type.

Reviewed By: prakashgayasen

Differential Revision: D86153749

fbshipit-source-id: d13d74ff931deefd1c774fb48402fc63382b337a