Set DOMElement::getAttribute return type to string
Summary:
getAttribute is currently untyped and can be a subtle source of bugs. This API is heavily depended on in the Video Delivery stack as it manipulates XML for client consumption

The API actually returns a string. See function declaration and implementation:

https://www.internalfb.com/code/fbsource/[e9ae5bc418b56330b97ee6f03e1733f23c5f2761]/fbcode/hphp/runtime/ext/domdocument/ext_domdocument.php?lines=881-882

https://www.internalfb.com/code/fbsource/[e9ae5bc418b56330b97ee6f03e1733f23c5f2761]/fbcode/hphp/runtime/ext/domdocument/ext_domdocument.cpp?lines=4101-4130

Switching the return type to string revealed a bunch of bugs in WWW mostly around using nullchecks to set a default value (it never returns null!), and implicit casting to enums without coerce/assert

Before making this change, I ran a local build of hh and fixed all the type errors introduced (see the parent diffs in the stack). This also included a search for dynamic invocations and mocking

Reviewed By: dlreeves

Differential Revision: D86128458

fbshipit-source-id: 7febceb065d822803cf0d0e6c695a2dc5bab67f7