Extend rank check
Summary: The current check looks for type parameters of higher rank when adding a bound to a type variable but not higher rank type variables. This omission allows higher rank generics to escape their scope. This diff extends the rank check to cover both cases and modifies the error message for the second case

Reviewed By: andrewjkennedy

Differential Revision: D86223152

fbshipit-source-id: eff5b85810db51f37d8953c0a5c9dc9d7a42c8f0