support dataclass in assert_thrift_almost_equal
Summary: The `dataclass` factory is commonly used, so assert_thrift_almost_equal needs to support it. However, it is not even `Iterable` so it needs some special casing. It's also a pain because every `dataclass` has no common (public) baseclass, so have to use the special `is_dataclass` function to check.

Reviewed By: yoney

Differential Revision: D86562527

fbshipit-source-id: 1cdc3f2fc1bd2222fa993908ce3566e01a98cea9