Persist rank check heuristic throughout typing
Summary:
The rank-check heuristic is used to avoid a traversal of a type to discover if it contains a type variable or parameter of a higher rank than the type variable to which we want to add it as an upper or lower bound.

Currently, the heuristic is carried in the subtype env. However, this information is lost during the transitive closure step and across calls to subtyping.

This diff moves the heuristic to the typing environment so that it is persisted throughout typing and makes additional checks during the transitivity step which give us better error messages - instead of multiple errors between a higher rank generic in the upper (lower) bound and the lower (upper) bounds of a type variable, we get a single error when we try to apply transitivity.

Reviewed By: andrewjkennedy

Differential Revision: D86589956

fbshipit-source-id: 33d98346c57a6e488ed2834109c95929f01758ff