Rewrite float in structured annotations
Summary:
Currently in metadata.thrift, float is always stored as double, but in schema.thrift we will convert it to float.

This conversion created a small discrepancy between metadata.thrift and schema.thrift.

This diff changed the structured annotations to have the same logic as schema.thrift.

Without this change, equality comparison will fail for annotations that contains float.

Reviewed By: sadroeck

Differential Revision: D86642320

fbshipit-source-id: 65c95a6e1be0bcdc736f2408102044d3c879b244