Propagate underlying source range when lowering DeprecatedUnvalidatedAnnotations
Summary:
When lowering the `thrift.DeprecatedUnvalidatedAnnotations` structured annotation to unstructured annotations, preserve the original source range of the original key/value pair in the unstructured annotation.

This ensures that emitting diagnostics (e.g. lints) against the lowered unstructured annotations will associate the diagnostic with the key/value pair that triggered it.

Reviewed By: sadroeck

Differential Revision: D86910193

fbshipit-source-id: 2d8be2543a399a8e198ebe9232a81562af330e40