Do not generate error for indexing on unknown shape field under dynamic assumptions
Summary: When checking under dynamic assumptions, which currently only is enabled under `--config everything_sdt=false`, we produce errors on indexing expressions `$x['a']` if `$x` was a value of type `dynamic` that has been refined using `Shapes::keyExists` to produce a type such as `shape('b' => int, ...)`. Under static assumptions, we retain the `dynamic &` under the refinement, but under dynamic assumptions, we unfortunately strip off the `dynamic`. This is one consequence of using "dynamic aware" subtyping when under dynamic assumptions. Until we revisit this decision, we can mitigate against this problem by not reporting an error: it's still sound, as the type we produce is the "open" type under `...`, just as it would be with `Shapes::idx`.

Differential Revision: D86879529

fbshipit-source-id: 9d06834ec97a754b5b22075bd7ff4835043733bc