Make exception_shared_string movable & assignable
Summary:
In my "rich error" work, it's quite helpful for this type to be regularly copyable/movable/assignable. It is basically a cheaper `std::string` for exception usage, and I didn't identify any C++-standard-related reasons it souldn't be movable & assignable.

The moved out state is `""`, not `nullptr`, because this backs `exception::what()` and that must always be safe to dereference, even if someone did mess up and end up reading a moved-out exception object.

Move support is C++20-only, since C++17 has a small & decreasing usage footprint, and it would take extra effort to support it.

# FAQ

## Why have `const` on the members and then cast it away?

I'm not sure, yfeldblum might know. To me, it seems like we no longer need the members to be const.

That said, it's harmless, so I'm leaving it for him to clean up.

## Should we also make it default-constructible, with the moved-out state equal to the default-constructed state?

Probably not. Having these be consistent is nice, but not required. And enforcing consistency comes with issues:

 - It is a code smell to construct exceptions with no `what()`, but default-constructibility encourages it.
 - Just like `std::string`, the type is not nullable, and should not become nullable. The default-constructed state must thus have a valid `what()` string.

Reviewed By: yfeldblum

Differential Revision: D86572375

fbshipit-source-id: ef302c20434e28713bac897505b7f39e1d8d12e8