Fix bug in elvis operator expressions
Summary:
If the LHS is truthy we return it so if it is readonly, the whole expression must also be treated readonly.

After writing the test, I was able to get Claude to fix this for me! (It seems' like the hphp/hack/CLAUDE.md file helped, but did seem to run into a lot of issues with being in the wrong directory I think when running commands; idk if you've noticed this madgen)

Reviewed By: dlreeves

Differential Revision: D86906025

fbshipit-source-id: 7ffd6be144d3b529fbaacc736e1a0acf731c67c9