fix get_all_files test
Summary:
This diff fixes the ` test_get_all_files` test of `hphp/hack/test/integration:edenfs_file_watcher_test ` which has started to become flakey.

The test has a list of files and directories it expects in the test repo in the beginning. Recently, a new folder can appear in a fresh `hg` repo. This diff adds this folder to the list of folders that we are expecting and can skip over.

Reviewed By: vassilmladenov

Differential Revision: D87113891

fbshipit-source-id: 97eb730e638cea022e2d42597c31104ccabf401e