Enusre repoOptionsHash is updated when reusing cached bytecode in unit-cache
Summary:
If we invalidate cached bytecode because of a change to `RepoOptionsFlags` (these flags can affect compilation) and then subsequently discover that the cached bytecode still matches we throw away the new compilation and preserve the original (this allows us to avoid invalidating translations in the TC and cached `Class` data).

The cache use code was previously grabbing the value of `repoOptionsHash` from the original `CachedFile` which in the case of a change to `.hhvmconfig.hdf` or `PACKAGES.toml` would now be incorrect. Thus altering these files will cause us to repeatedly spin through the unit cache recompiling files over and over again.

The new test demonstrates how this behavior can go awry, in P2040523862 you can see the output of this test before the fixes here.

Differential Revision: D87127599

fbshipit-source-id: e37ca051eb416fcc11c101a91478871321679f4a