IoUringProvidedBufferRing: pack and optimise size
Summary: Pack the struct tightly to reduce gaps. Shuffle member fields around and put the hottest read/write fields in the first cacheline. This is assuming that freeing is on the EVB thread (Ucache). Will need to profile and investigate if future users do not use Thrift EB mode and move work onto CPU thread pools.

Reviewed By: vishwanath1306

Differential Revision: D87154921

fbshipit-source-id: 68be77f5a43eb1191270ef073455db183cd6e785