relax reified generics debug assert for type structure interning
Summary:
Even if two type structures represent the same type, they may differ in the way they
were constructed due to how constraints are resolved in inheritance chains vs the default
behavior of ?T.

See the introduced test for an example of the problem.

In practice, we don't think relaxing the check has any downstream runtime issues.
The real solution is to rip all of this machinery out, but that's more involved so let's ensure
we avoid crashes when we encounter this case.

Reviewed By: paulbiss

Differential Revision: D87232450

fbshipit-source-id: 6612f4238386af7349d725f4623cb9f24a905a7a