Simplify treatment of parent::__construct and fix issues with like types
Summary:
The logic for checking `parent::__construct` calls is shared with that for checking `new`. It's broken in at least a couple of ways:
* There is no check that a constructor exists in the parent (T38050584)
* It doesn't properly support like-type calling i.e. treating the constructor as "supporting dynamic"
Also, the logic is kind of nonsense: for a base class that has generics e.g. `Base<TB>` we generate fresh type variables for the generics, do all the argument checking, and then after the fact check that `Base<#1> = Base<TB>`.

It's much simpler instead to treat `parent::__construct` just like a normal call. We don't need the nonsense of fresh type variables, e.g. if we have `Derived<T> extends Base<vec<T>>` then we're simply checking parameter types that involve `T` in scope. But in order to enable like-type calling, we must make sure that `supportdyn` is wrapped around the function type of the constructor, just as with methods.

Note that we already check constructors under both static and dynamic assumptions, so this "use-site" `supportdyn` treatment now fits with the "definition-site" checking of the constructor.

Differential Revision: D87070529

fbshipit-source-id: d37417f699d57fcacf27de9d37cc1d071a0a24ab