add heuristics based on classes enclosing roots
Summary:
This diff adds the following heuristics to `hh --find-my-tests`. For any root `r` we now look for references to the enclosing class `C` and do the following:
- If `C::class` or `new C(...)` appears in the definition of some symbol `q`, and `q` is not inside a test file, we add `q` to our graph to be explored further.
- If `C` is referenced in any way in a test file, we select that test file. Note that this is fairly broad: It goes beyond the cases `C::class` and `new C(...)` mentioned above, but is for example also triggered by `C` being used as a parameter type.

Practically speaking, this means that we now have classes as a new type of nodes in our graph. Since the only previously supported type of nodes were methods, `q` and `r` will always be methods in the text above when running the algorithm today. But this will change in the future.

Drive-by changes:
- Fixed the BUCK definition of the test to use `$(exe ...)` instead of `$(location ...)`.
- Added lots of logging to `ServerFindMyRefs`, using `Hh_logger` category "FindMyTests"

Reviewed By: madgen

Differential Revision: D87095977

fbshipit-source-id: ce66a7e24c061adc8754a460a34f6b84e63d08aa