Move return outside foreach scope
Summary:
Foreach loops are wrapped inside an implicit try catch block so that the iterators can be freed in the event of an exception in the loop body. Instead of this, emit the return outside the scope of the try catch. This is required for the next diff where its possible to double free iterators after the VerifyRetTypeC bytecode is merged with RetC.

Previously:
```
try:
    ...
    VerifyRetTypeC
    IterFree 0
    RetC
    ...
catch:
   IterFree 0
   throw
```

Now:
```
try:
    ...
    IterFree0
    Jmp L0
    ...
catch:
    IterFree 0
    throw
L0:
   ...
   VerifyRetTypeC
   RetC
```

Reviewed By: jano

Differential Revision: D85533781

fbshipit-source-id: 4b178cd2c9e0a5a086368399a6b7322f1eccb477