Turn HackLint5811 into advice because it's not really an error
Summary:
Error means "This is something critical and is an issue if landed." See https://www.internalfb.com/wiki/Hack_Developer_Guide/Linters/aurora-the-hack-linter/writing-aurora-lints/

This is not that.

Also AI is overreacting to it, see https://fb.workplace.com/groups/wfi/posts/30068701396085089

Downgrading to "advice" lowers the penalty to CQS which is what drives most of the AI codemods while still retaining the lint for analysis and potential improvements. It's fine if this is upgraded later provided it doesn't cause issues / has maintenance and support around it, but as-is people are hurting the site by trying to fix this.

Reviewed By: WizKid

Differential Revision: D87392889

fbshipit-source-id: 0b05d0f80fae86545872315ab58e2bbbf7bbceb5