Use optimised forms of class release functions where possible
In makeDtorCall we call a release function of the form

void release(ObjectData *obj, Class *cl)

but in some cases we load the second class pointer argument
directly from the object pointer. However, we can actually
invoke a more optimised version of the release function:

void release2(ObjectData *obj)

and let this release function do the work of extracting the
class pointer.

Doing so helps to reduce the total size of the Jit code cache
by removing an associated load for every call. It may slightly
increase the total instruction execution count by adding an
extra level of indirection, but I think it's worth it.