Fix speculation bug in HHBBC
Summary:
When we're speculating in a block, we can't do any reductions
which involve rewinding the stack or popping/pushing different
amounts. During speculation we track stack push/pops manually so we
can determine if the speculated block leaves something on the stack.

So, if we rewind the stack or change a bytecode's push or pops, the
tracking will get out of sync. This can result in miss-optimizations,
as it means we may successfully speculate when we shouldn't. The
attached test case is an example.

Add a helper function which designates whether it's safe to rewind the
stack. Use that helper before rewinding.

While I'm at it, change some FTRACE to ITRACE so that their output is
properly aligned.

Reviewed By: paulbiss

Differential Revision: D87418319

fbshipit-source-id: 38d8a67285e7c7b12889d34c3c1290dc1d910b30