use reported maxrss from workers to inform how many we schedule
Summary:
For large workloads and large machines, SubprocessImpl may overschedule workers - since the
pre-existing tuning was one subprocess per thread, jobs which have memory usage that's more than
RAM / #cores would hit memory limits.

This diff introduces the idea of a SubprocessScheduler to address this problem. Idea is simple -
SubprocessScheduler gets initialized with a memory cap, and before scheduling any managed PID, the caller calls `tryAcquire` and `registerSubprocess` (the latter to enable RSS profiling for the processes), and releases
the PID with the actual usage to free that amount of space for the scheduler to allow more work. `tryAcquire` only allows scheduling work if either (a) no work is scheduled or (b) the current usage estimate + new job's estimate would remain under the cap.

The initial memory estimates are set high to avoid an unsampled new job to OOM the machine, and gives the
scheduler time to learn the actual behavior online.

Reviewed By: ricklavoie

Differential Revision: D84875099

fbshipit-source-id: 47fca8efc5ded7cd70c4c67bcbf55dcbe1ee280e