compress files to disk for SubprocessClient
Summary:
ZSTD is fast (50MB/s+ decompression/compression), and disk reads are slow. We've observed that large workloads incur significant disk read/write load, which ends up slowing HHBBC down significantly.

This diff changes how blobs are stored and read - blob references now use `m_extra` as a (uint32, uint32) pair that marks an (offset, compressed_length) index into a blobfile. Blobs are compressed when writing/decompressed when reading now, and this is handled by breaking up the previously existing FD abstraction into `BlobReader` and `BlobWriter`'s which  compress/decompress transparently.

We also add support for an optional compression dictionary. The way modern compression algorithms work is by learning the shape of data online as they read the  start of the stream. This works well for large files, but since a lot of our blobs are small, the compressor won't have time to learn the structure of the data well enough to apply advanced rules. Compression dicts let you pre-encode data by training the compressor on a set of samples. We could experiment with compression dictionaries that encode the structure of data so that this learning phase can be skipped and we end up with better compression ratios without decoding/encoding expense.

Reviewed By: ricklavoie

Differential Revision: D85068903

fbshipit-source-id: 46b3dd523483334be37aed90bc9497adb526e531