Fix issue where symlinked directories would cause wrong permissions
Summary:
Context:
https://fb.workplace.com/groups/watchman.users/permalink/25162561930022020/
JestE2E was seeing issues in sandcastle where watchman get-sockname was failing on permission errors.

After investigating I found that the issue was there was an intermediate step where watchman tries to set the group ACL, but was using lstat instead of stat.
Lstat does not follow symlinks, so it was using the symlink's 777 instead of the target's permission as a base when setting the group permissions.
Immediately after, watchman sets the target's permissions to 750, which causes a race condition that occurs when you have multiple instances trying to do this at the same time, as happens in jest's call to static resources.

Reviewed By: genevievehelsel

Differential Revision: D87464522

fbshipit-source-id: e060e41c586eb35b37485c79c997f25ee3341212