IoUringZeroCopyBufferPool: implement import and export
Summary:
Export is done from the owning io_uring instance and returns an ExportHandle.

This move-only handle can be moved to other EVBs, and then imported via importHandle() to create a new IoUringZeroCopyBufferPool that can be used in that EVB.

On the get side, the only synchronisation is an atomic counter. The work is naturally going to be non-overlapping e.g. buffer indices.

On the release side, everything will contend on a folly::DistributedMutex to be able to write to the refill queue. If this is sus later on, we can revisit with local return + batched refill or maybe kernel side changes for NAPI to round-robin multiple refill queues.

Reviewed By: vishwanath1306

Differential Revision: D87184181

fbshipit-source-id: cd1f4a1a5311b08bd0e8be3dfb3226f6f81f959e