unique_hash_key
Summary:
`unique_hash_key` is a fixed-size and unique hashtable key for which hashing is cheap. Holds a cryptographic hash of the arguments with which it is constructed - cryptographic hashes are fixed-size and guarantee uniqueness. Hashing for a hashtable just reads the first word.

Extracted and generalized from `RegexMatchCacheKey`, which is used in `RegexMatchCache`.

Optimized for performance. Keeps OpenSSL's SHA256 but switches to a keyed construction and avoids allocations. Offers two additional variants: BLAKE3 which is cryptographically strong, and XXH3 which is faster but not cryptographically strong.

Allows construction from a pack of data items where all of the data items contribute to the final hash, not just from a single string key. Currently supports non-`bool` integer types and `std::string_view` data items. Hash keys constructed from different arguments are only expected to give different results when the different arguments have the same shape, i.e., the same hash algorithm and packs with element-wise matching types.

Reviewed By: lexprfuncall

Differential Revision: D85528861

fbshipit-source-id: fd5699cd5f437c84ff1aeb6e749065ad9ea6b835