Disable Blake3 on Windows in UniqueHashKey.cpp
Summary: Stuff depending on this library (`buck build --flagfile fbcode//mode/opt-win fbcode//common/stats/py:servicedata/servicedata`) fails to build on Windows because the Blake3 header is available, but the library is not. I tried reverting D85528861 entirely but there's already a diff that depends on it, and other on the way. Instead I'm going to try to disable Blake3 for Windows.

Reviewed By: khoi0107

Differential Revision: D87731375

fbshipit-source-id: b023dd70416b2f22de2ecaada5e7dc8bc8ff9ac8