Fix loosen_array_staticness bug in HHBBC
Summary:
If an array type is static, then implicitly it's contents must be
static too. We do not use a specialization for this case, as it's
strictly redundant. loosen_array_staticness "loosens" an array type so
that it is no longer static. This conversion is not done recursively,
so the inner array contents are unchanged (this is meant to model the
effects of COW).

However, loosen_array_staticness simply manipulates the trep bits of
the type. This means that, for example, a TSDictN becomes just a
TDictN. This loses the implicit knowledge about the array contents
being static! This may not seem important since we're widening the
type anyways, but it's necessary to keep this information to ensure
monotonicity.

Have loosen_array_staticness check for this situation, and apply a
specialization to the loosened array to preserve staticness
information about array contents.

Also do a little bit of cleanup while I'm at it.

Reviewed By: mdko

Differential Revision: D87744013

fbshipit-source-id: 0876e8854d25584de09ef97411a0a9eba963411f