update named parameters completion test so it reflects reality
Summary:
## What

Update tests for completions for named parameters to use methods rather than top-level functions

## Why

In prod, we do not complete parameters for top-level functions, so the test was unrealistic.

## More background

You can see the real behavior here: we never complete parameters for top-level functions, but do for methods. A comment says that this is for performance reasons.

 https://pxl.cl/8vmhH

## Follow-on work

- Update our test suite settings so they are closer to prod to avoid misleading test snapshots like what we had before: D87784064
- Fix the snippets for named parameter completions, since the `tab` behavior is currently wrong: D87794413

Differential Revision: D87784065

fbshipit-source-id: 10a1f1c9a31d2cf4eed56b33ff66cf6986066fa0