correct named argument completions
Summary:
### What

Fix completions for calls to functions with named parameters.

With this change:

https://pxl.cl/8vmrW

Without this change, when the user `tab`s to a named argument, the selection covers the name as well, so the name of the named argument is deleted when the user types (bad)

 https://pxl.cl/8vms4

## Additional testing

I also added some tests that show the behavior with optional parameters. We do not show optional parameters in completions. This matches the existing behavior for positional optional parameters.

## Reference

[Language Server Protocol Snippet Syntax (docs)](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#snippet_syntax)

Differential Revision: D87794413

fbshipit-source-id: fd12c037fa88831aa58f4f1b051d0c70f6539d0b