Unify null-safe pipe and null-safe member selection internal representation
Summary:
Changes

```
type og_null_flavor = Ast_defs.og_null_flavor =
  | OG_nullthrows
  | OG_nullsafe
```

to

```
type operator_null_flavor = Ast_defs.operator_null_flavor =
  | Regular
  | Nullsafe
```

to make sure that it is applicable not to just `ObjectGet` but to `Pipe`  as well. Using this variant for null-safe pipe implementation allowed to get rid of a less descriptive boolean flag that was used to distinguish between regular and null-safe pipes.

This is purely implementation changes so it is no-op from use's point of view.

Reviewed By: mheiber

Differential Revision: D86202098

fbshipit-source-id: 267b8ec1a1ffe88d1ef75a9a4f439345ee6a38d9