- fix error: declaration shadows a local variable
Summary:
I noticed:
  buck2 build fbsource//arvr/libraries/Shaka:shaka_common

fails,  with shadowing errors

this is because its compiling xplat/folly/futures:core with
    "-Wshadow-all",
    "-Werror=shadow-uncaptured-local",
and folly/futures has several place where local variables have the same name as parameters or member variables.

I opted to use the _2 postfix syntax, as I noticed it was already used elsewhere in the folly code. Using a single _ does not work when this syntax is used alread for members vs parameters

Reviewed By: yfeldblum

Differential Revision: D87841581

fbshipit-source-id: 47217187ad677b14fcbee5b2c45ce1208b8046e8