Fix prefix match bug
Summary:
File prefix matches support `**` (globstar); the current implementation has a bug which means that when we have a required exact segment match which fails, it gets matched by the globstar.

This diff fixes the bug and adds more 'bad' units test for globstar patterns. It also hoists the pattern matching code out of the prefix / exact matches so that they share an implementation with greater test coverage.

Finally, this code has been rewritten from CPS to ANF to make it easier to read.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D87923876

fbshipit-source-id: e6575bde04688e7ed191d203d2b38de85ec47bdc