StripedEDFThreadPoolExecutor
Summary:
`EDFThreadPoolExecutor` is not scalable: in part this is due to design flaws such as allocating and deallocating queue entries in the critical section, but for the largest part this is fundamentally due to trying to maintain a global priority queue (there is some attempt at sharding, but it does not look effective).

This diff introduces an alternative implementation of EDF, which keeps one priority queue per LLC. On servers with one LLC, this makes no difference, while on larger servers we trade off some accuracy in picking the next task for reduced contention. With large core numbers, the impact on prioritization should be negligible.

The queue implementation is also very different: instead of a set of binary heaps, we use just a single intrusive heap, protected by a `DistributedMutex`.

Reviewed By: yfeldblum

Differential Revision: D87927531

fbshipit-source-id: 753a27aa50ef8564ff278ab976c7a9d8295ae7ae