Convert the string type into a class
Summary:
Switch the representation of the `string` type from a `Tprim` to a class in `string.hhi`. This is a straightforward refactor in most of the code. However, there are a few places where large match expression need to get moved around non-locally, and those are fiddly: `typing_equality_check.ml`, and the disjointness code in `typing_subtype.ml`.

We aim to make this change as unobservable as possible
  - `typing_json.ml` prints the string type as a primitive rather than a class
  - error messages try to not change (e.g., not printing HH\string, even where it would print HH\vec)
  - runtime messages from reflection stay the same (see the .rs files)

There is some churn in .exp files due to `string` not sorting exactly the same in unions.

Later, when we want to make the new representation more visible, we can just search for `Classes.cString` for places where some special treatment is applied.

Reviewed By: andrewjkennedy

Differential Revision: D86792448

fbshipit-source-id: 35c2be08e0125ca10f3f33cb722b46b6647b3813