Do not attempt to check constraints for instantiated types if the arity is wrong
Summary:
We would like to avoid any generation of `Tany` without the type-checker.

One place it is generated is when checking constraints on generic arguments to types that are ill-formed with respect to arity. For example, if we have
```
class C<T1 as I<T2>, T2 as J> { ... }
```
and the use of `C` such as
```
function foo(C<D> $x): void { }
```
then we will generate a substitution `{T1:D, T2:Tany}` in order to check `D <: I<Tany>`.

This is unnecessary: we already emit an arity error. There is no point in continuing with a nonsensical constraint check.

Reviewed By: geralt-encore

Differential Revision: D88005980

fbshipit-source-id: 82d99da289e423d056f991297a4fe7c2be1c20f2