DynamicValue 9/n: iterators
Summary:
Enables range-based iteration and STL compatibility
Note that due to type erasure the loop variable is a variable rather than a reference:
     for (auto [keyRef, valueRef] : constMap) { ... }
Likewise, the iterator implementation has to be polymorphic, but we want to avoid requiring allocation for const access. As a result we introduce `SmallBuffer` to store the concrete implementation inline with appropriate safeguards.

Reviewed By: praihan

Differential Revision: D87574181

fbshipit-source-id: d0a0b25ea79777550790a5b756c24c9d97874232