DynamicValue 1/n: skeleton
Summary:
DynamicValue is the type-safe successor to protocol value.
It uses TypeSystem to tag values with their runtime schemas and provide checked access to and operations on the contents.
The guiding principle is to provide value semantics.
DynamicRef is introduced to provide mutable views into containers/structs in later diffs.

Reviewed By: praihan

Differential Revision: D85101166

fbshipit-source-id: 57e39f6d4cd23cbcef77e82cad97318f5af2546d