test extended reasons for named parameters
Summary:
Show current behavior of extended reasons with named parameters.

I actually think this isn't too bad and is OK to leave as is.

Our explanation mentions arguments by name and then points to where the arguments are. It's not wrong, and the interesting part to look at is the actual code locations:

https://www.internalfb.com/diff/D87792055?dst_version_fbid=1819337738690389&transaction_fbid=839992801949544

## Potential future work

Whether AIs will get confused by referring to named parameters by index in extended reasons messages is an empirical question we could answer if/when there's big adoption of named parameters and it seems worth it to measure.

Note that if we end up diving into which wording works best for AIs or humans, we may also want to look into whether deconflating "argument" and "parameter" would help (see same example linked above).

Reviewed By: enetsee

Differential Revision: D87792055

fbshipit-source-id: 7f19913ffd9490e91df1a1fcd10d6142b5e7a375