Set SOCK_CLOEXEC for debugger sockets
It's possible to configure HHVM access logs to be sent to a subprocess
instead of a file, like so:
```
Log {
	Access {
		* {
			File = |/usr/bin/some_processor.sh | tee -a /tmp/hhvm-access.log
		}
	}
}
```
Under the hood, this is backed by a `popen(3)` syscall in
`ClassicWriter`, which ultimately creates a pipe, forks and invokes a
shell. If HHVM is running with the debugger enabled
(`hhvm.debugger.vs_debug_enable=1`), this subprocess also ends up
inheriting the file descriptor backing the listening socket of the
debugger. If a connected debugger client then disconnects,
the debug server will then restart and try to open a new socket, but fail
because the FD remains open in the child process.

So, set `SOCK_CLOEXEC` on these sockets so that they get closed in child
processes.